re-use fast_inhous resampler test suite for hifi_rubato
