+++ src/conversions/sample_rate.rs
@@ -1,3 +1,4 @@
 pub(crate) mod fast_inhouse;
 pub(crate) mod hifi_rubato;
-
+#[cfg(test)]
+mod test;
